requirements_win.txt
nepali-datetime
openpyxl
reportlab
pywin32
ttkbootstrap
mysql-connector-python